rf: move Node tests from ManageJenkinsTest -> new class NodeTest
