+++ src/test/java/school/redrover/ManageJenkinsTest.java
@@ -9,14 +9,11 @@ import school.redrover.page.home.HomePage;
 import school.redrover.page.systemConfiguration.SystemPage;
 import school.redrover.runner.BaseTest;
 import school.redrover.runner.ProjectUtils;
-import school.redrover.runner.TestUtils;
 
 import java.util.List;
 
 public class ManageJenkinsTest extends BaseTest {
 
-    private static final String MY_NODE_NAME = "My name of node";
-    private static final String NODE_NAME = "NewNodeName";
     private static final String FULL_USER_NAME = "Ivan Petrov";
 
     @DataProvider
@@ -281,63 +278,6 @@ public class ManageJenkinsTest extends BaseTest {
                 .getTitle();
 
         Assert.assertEquals(pageTitle, "Thread Dump");
-    }
-
-    @Test
-    public void testCreateNewAgent() {
-
-        List<String> nodeItemList = new HomePage(getDriver())
-                .openManageJenkinsPage()
-                .clickNodesButton()
-                .clickButtonNewNode()
-                .enterNodeName(NODE_NAME)
-                .selectPermanentAgent()
-                .clickButtonCreate()
-                .clickButtonSave()
-                .getNodeList();
-
-        Assert.assertListContainsObject(nodeItemList, NODE_NAME, "List not contain new node");
-    }
-
-    @Test
-    public void testCreationAndDisplayNewNameOnNodesPage() {
-        List<String> itemNoteList = new HomePage(getDriver())
-                .openManageJenkinsPage()
-                .clickNodesButton()
-                .clickButtonNewNode()
-                .enterNodeName(MY_NODE_NAME)
-                .selectPermanentAgent()
-                .clickButtonCreate()
-                .clickButtonSave()
-                .getNodeList();
-
-        Assert.assertListContainsObject(itemNoteList, MY_NODE_NAME, MY_NODE_NAME);
-    }
-
-    @Test()
-    public void testDeleteAgent() {
-        TestUtils.createNode(getDriver(),NODE_NAME);
-
-        List<String> nodeNameAfterList = new HomePage(getDriver())
-                .openNodeFromBuildExecutorStatusBlock(NODE_NAME)
-                .clickDeleteButtonSidebarAndConfirm()
-                .gotoHomePage()
-                .getNodeNameList();
-
-        Assert.assertEquals(nodeNameAfterList.size(), 0);
-    }
-
-    @Test
-    public void testDeleteAgentByChevronFromBuildExecutorsStatus() {
-        TestUtils.createNode(getDriver(), NODE_NAME);
-
-        List<String> nodeNameAfterList = new HomePage(getDriver())
-                .selectDeleteAgentFromBuildDropdownAndClickYes(NODE_NAME)
-                .gotoHomePage()
-                .getNodeNameList();
-
-        Assert.assertEquals(nodeNameAfterList.size(), 0);
-        Assert.assertListNotContainsObject(nodeNameAfterList, NODE_NAME,"List contains node");
     }
 
     @Test
